# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Generic SNAP/ASPI-compliant payment gateway API: B2B access\ntoken issuance, admin client onboarding, signature-generation\nutilities, and Virtual Account (transfer-VA) transaction and\nmerchant-dashboard endpoints.",
        "title": "SNAP Payment Integration Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/clients": {
            "post": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Creates a new client_app trust-anchor record and, if publicKeyPem is supplied, its first client_key in the same call. State-changing: creates persistent client onboarding records.",
                "tags": [
                    "Admin Client Management"
                ],
                "summary": "Register a new B2B client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Static admin API key (ADMIN_API_KEY)",
                        "name": "X-Admin-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Client registration payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RegisterClientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload, missing clientId/clientName/keyId, or invalid publicKeyPem",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Invalid or missing X-Admin-API-Key header",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "503": {
                        "description": "Admin API disabled (ADMIN_API_KEY not set)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/clients/{clientId}/keys": {
            "post": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Registers an additional active public key for an existing client (e.g. key rotation). State-changing: creates a persistent client_key record.",
                "tags": [
                    "Admin Client Management"
                ],
                "summary": "Add a client key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Static admin API key (ADMIN_API_KEY)",
                        "name": "X-Admin-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client identifier",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Client key payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AddClientKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "400": {
                        "description": "Missing clientId, missing keyId/publicKeyPem, or invalid publicKeyPem",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Invalid or missing X-Admin-API-Key header",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "503": {
                        "description": "Admin API disabled (ADMIN_API_KEY not set)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/clients/{clientId}/keys/{keyId}": {
            "delete": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Revokes (deactivates) a client's public key so it can no longer be used to verify signatures. State-changing: updates a persistent client_key record.",
                "tags": [
                    "Admin Client Management"
                ],
                "summary": "Revoke a client key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Static admin API key (ADMIN_API_KEY)",
                        "name": "X-Admin-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client identifier",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key identifier to revoke",
                        "name": "keyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "400": {
                        "description": "Missing clientId or keyId path parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Invalid or missing X-Admin-API-Key header",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.clientResponse"
                        }
                    },
                    "503": {
                        "description": "Admin API disabled (ADMIN_API_KEY not set)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AddClientKeyRequest": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "keyId": {
                    "type": "string"
                },
                "publicKeyPem": {
                    "type": "string"
                }
            }
        },
...
        "domain.ClientApp": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "client_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ClientStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.ClientKey": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "key_id": {
                    "type": "string"
                },
                "public_key_pem": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.ClientStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "REVOKED",
                "SUSPENDED"
            ],
            "x-enum-varnames": [
                "ClientStatusActive",
                "ClientStatusRevoked",
                "ClientStatusSuspended"
            ]
        },
...
        "domain.RegisterClientRequest": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "clientId": {
                    "type": "string"
                },
                "clientName": {
                    "type": "string"
                },
                "keyId": {
                    "type": "string"
                },
                "publicKeyPem": {
                    "type": "string"
                }
            }
        },
...
        "handler.clientResponse": {
            "type": "object",
            "properties": {
                "client": {
                    "$ref": "#/definitions/domain.ClientApp"
                },
                "key": {
                    "$ref": "#/definitions/domain.ClientKey"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminToken": {
            "description": "Static admin API key configured via the ADMIN_API_KEY environment variable. Required for all /admin/* endpoints.",
            "type": "apiKey",
            "name": "X-Admin-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Bearer access token issued by POST /openapi/v1.0/access-token/b2b, e.g. \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SnapClientKey": {
            "description": "Client identifier issued by the admin during onboarding. Required on the B2B access-token endpoint.",
            "type": "apiKey",
            "name": "X-CLIENT-KEY",
            "in": "header"
        },
        "SnapSignature": {
            "description": "Request signature. Compute via POST /api/v1/utilities/signature-auth (asymmetric, access-token endpoint) or POST /api/v1/utilities/signature-service (symmetric, transaction endpoints).",
            "type": "apiKey",
            "name": "X-SIGNATURE",
            "in": "header"
        },
        "SnapTimestamp": {
            "description": "Request timestamp in ISO 8601 format (e.g. 2026-07-24T10:00:00+07:00).",
            "type": "apiKey",
            "name": "X-TIMESTAMP",
            "in": "header"
        }
    }
}